requirements.txt
attrs==23.2.0
boto3==1.34.69
botocore==1.34.69
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
isodate==0.6.1
jmespath==1.0.1
lxml==5.1.0
numpy==1.26.4
pandas==2.2.1
platformdirs==4.2.0
psycopg2-binary==2.9.9
PyGreSQL==6.0
pyspark==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
requests-file==2.0.0
requests-toolbelt==1.0.0
s3transfer==0.10.1
six==1.16.0
tzdata==2024.1
urllib3==1.26.18
zeep==4.2.1